# parking/api/swagger/swagger.yaml
swagger: '2.0'
info:
  version: 0.0.1
  title: Parking api
host: 'localhost:10010'
basePath: /v1
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /admin/login:
    post:
      tags:
        - Auth
      x-swagger-router-controller: auth
      operationId: loginAdmin
      summary: Valida las credenciales del admin
      description: ''
      produces:
        - application/json
      parameters:
        - in: body
          name: credenciales
          description: Credenciales del usuario
          required: true
          schema:
            $ref: '#/definitions/credenciales'
      responses:
        '200':
          description: Operación exitosa
          schema:
            $ref: '#/definitions/dataLoginAdmin'
        '400':
          description: Solicitud incorrecta
          schema:
            $ref: '#/definitions/mensajeModel'
        '401':
          description: Credenciales incorrectas
          schema:
            $ref: '#/definitions/mensajeModel'
        '500':
          description: Error en sistema
          schema:
            $ref: '#/definitions/mensajeModel'
  /admin/loginF:
    post:
      tags:
        - Auth
      x-swagger-router-controller: auth
      operationId: loginAdminF
      summary: Valida las credenciales del admin financiero
      description: ''
      produces:
        - application/json
      parameters:
        - in: body
          name: credenciales
          description: Credenciales del usuario
          required: true
          schema:
            $ref: '#/definitions/credenciales'
      responses:
        '200':
          description: Operación exitosa
          schema:
            $ref: '#/definitions/dataLoginAdminF'
        '400':
          description: Solicitud incorrecta
          schema:
            $ref: '#/definitions/mensajeModel'
        '401':
          description: Credenciales incorrectas
          schema:
            $ref: '#/definitions/mensajeModel'
        '500':
          description: Error en sistema
          schema:
            $ref: '#/definitions/mensajeModel'
  /admin/pagos:
    post:
      tags:
        - Reserva
      x-swagger-router-controller: reserva
      operationId: reservaPagadas
      summary: Busca todos los pagos del sistema
      description: ''
      produces:
        - application/json
      parameters:
        - in: header
          name: api_key
          description: ''
          required: true
          type: string
      responses:
        '200':
          description: Operación exitosa
          schema:
            $ref: '#/definitions/pagosUsuarios'
        '400':
          description: Solicitud incorrecta
          schema:
            $ref: '#/definitions/mensajeModel'
        '401':
          description: Credenciales incorrectas
          schema:
            $ref: '#/definitions/mensajeModel'
        '500':
          description: Error en sistema
          schema:
            $ref: '#/definitions/mensajeModel'
      security:
        - api_key: []
  /admin/estadopagos:
    post:
      tags:
        - Reserva
      x-swagger-router-controller: reserva
      operationId: reservaEstados
      summary: Obtiene un resumen de los estados de los pagos
      description: ''
      produces:
        - application/json
      parameters:
        - in: header
          name: api_key
          description: ''
          required: true
          type: string
      responses:
        '200':
          description: Operación exitosa
          schema:
            $ref: '#/definitions/estadoPagos'
        '400':
          description: Solicitud incorrecta
          schema:
            $ref: '#/definitions/mensajeModel'
        '401':
          description: Credenciales incorrectas
          schema:
            $ref: '#/definitions/mensajeModel'
        '500':
          description: Error en sistema
          schema:
            $ref: '#/definitions/mensajeModel'
      security:
        - api_key: []
...
  /auto:
    post:
      tags:
        - Auto
        - User
      x-swagger-router-controller: auto
      operationId: createAuto
      summary: Crea un auto para el usuario
      description: ''
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: Informacion del auto
          schema:
            $ref: '#/definitions/auto'
        - in: header
          name: api_key
          description: ''
          required: true
          type: string
      responses:
        '200':
          description: Operación exitosa
          schema:
            $ref: '#/definitions/mensajeModel'
        '400':
          description: Solicitud incorrecta
          schema:
            $ref: '#/definitions/mensajeModel'
        '403':
          description: Credenciales incorrectas
          schema:
            $ref: '#/definitions/mensajeModel'
        '500':
          description: Error en sistema
          schema:
            $ref: '#/definitions/mensajeModel'
      security:
        - api_key: []
    put:
      tags:
        - Auto
        - User
      x-swagger-router-controller: auto
      operationId: updateAuto
      summary: Actualiza el auto del usuario
      description: ''
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: Informacion del auto
          schema:
            $ref: '#/definitions/auto'
        - in: header
          name: api_key
          description: ''
          required: true
          type: string
      responses:
        '200':
          description: Operación exitosa
          schema:
            $ref: '#/definitions/mensajeModel'
        '400':
          description: Solicitud incorrecta
          schema:
            $ref: '#/definitions/mensajeModel'
        '403':
          description: Credenciales incorrectas
          schema:
            $ref: '#/definitions/mensajeModel'
        '500':
          description: Error en sistema
          schema:
            $ref: '#/definitions/mensajeModel'
      security:
        - api_key: []
  /auto/{id}:
    delete:
      tags:
        - Auto
        - User
      x-swagger-router-controller: auto
      operationId: deleteAuto
      summary: Elimina el auto del usuario
      description: ''
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          description: ID del auto el cual sera eliminado
          required: true
          type: string
        - in: header
          name: api_key
          description: ''
          required: true
          type: string
      responses:
        '200':
          description: Operación exitosa
          schema:
            $ref: '#/definitions/mensajeModel'
        '400':
          description: Solicitud incorrecta
          schema:
            $ref: '#/definitions/mensajeModel'
        '403':
          description: Credenciales incorrectas
          schema:
            $ref: '#/definitions/mensajeModel'
        '500':
          description: Error en sistema
          schema:
            $ref: '#/definitions/mensajeModel'
      security:
        - api_key: []
  /reserva/{id}:
    get:
      tags:
        - Reserva
        - Auto
        - User
      x-swagger-router-controller: reserva
      operationId: detalleReserva
      summary: Muestra el detalle de una reserva
      description: ''
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          description: ID de la reserva
          required: true
          type: string
        - in: header
          name: api_key
          description: ''
          required: true
          type: string
      responses:
        '200':
          description: Operación exitosa
          schema:
            $ref: '#/definitions/reserva'
        '400':
          description: Solicitud incorrecta
          schema:
            $ref: '#/definitions/mensajeModel'
        '403':
          description: Credenciales incorrectas
          schema:
            $ref: '#/definitions/mensajeModel'
        '500':
          description: Error en sistema
          schema:
            $ref: '#/definitions/mensajeModel'
      security:
        - api_key: []
    delete:
      tags:
        - Reserva
      x-swagger-router-controller: reserva
      operationId: deleteReserva
      summary: Elimina una reserva
      description: ''
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          description: ID de la reserva
          required: true
          type: string
        - in: header
          name: api_key
          description: ''
          required: true
          type: string
      responses:
        '200':
          description: Operación exitosa
          schema:
            $ref: '#/definitions/mensajeModel'
        '400':
          description: Solicitud incorrecta
          schema:
            $ref: '#/definitions/mensajeModel'
        '403':
          description: Credenciales incorrectas
          schema:
            $ref: '#/definitions/mensajeModel'
        '500':
          description: Error en sistema
          schema:
            $ref: '#/definitions/mensajeModel'
      security:
        - api_key: []
  /reserva:
    post:
      tags:
        - Reserva
        - Auto
      x-swagger-router-controller: reserva
      operationId: createReserva
      summary: Crea una nueva reserva
      description: ''
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: Crea un usuario
          schema:
            $ref: '#/definitions/reserva'
        - in: header
          name: api_key
          description: ''
          required: true
          type: string
      responses:
        '200':
          description: Operación exitosa
          schema:
            $ref: '#/definitions/mensajeModel'
        '400':
          description: Solicitud incorrecta
          schema:
            $ref: '#/definitions/mensajeModel'
        '403':
          description: Credenciales incorrectas
          schema:
            $ref: '#/definitions/mensajeModel'
        '500':
          description: Error en sistema
          schema:
            $ref: '#/definitions/mensajeModel'
      security:
        - api_key: []
  /estacionamiento:
    post:
      tags:
        - Estacionamiento
      x-swagger-router-controller: estacionamiento
      operationId: createEstacionamiento
      summary: Crea un estacionamiento para el dueño
      description: ''
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: Informacion del estacionamiento
          schema:
            $ref: '#/definitions/estacionamiento'
        - in: header
          name: api_key
          description: ''
          required: true
          type: string
      responses:
        '200':
          description: Operación exitosa
          schema:
            $ref: '#/definitions/mensajeModel'
        '400':
          description: Solicitud incorrecta
          schema:
            $ref: '#/definitions/mensajeModel'
        '403':
          description: Credenciales incorrectas
          schema:
            $ref: '#/definitions/mensajeModel'
        '500':
          description: Error en sistema
          schema:
            $ref: '#/definitions/mensajeModel'
      security:
        - api_key: []
  /estacionamiento/{id}/horario:
    get:
      tags:
        - Estacionamiento
        - Horario
      x-swagger-router-controller: estacionamiento
      operationId: findHorarios
      summary: obtiene los horarios del estacianamiento
      description: Permite la obtener todos los auto  del estacionamiento
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          description: ID del estacionamiento el cual sera buscado
          required: true
          type: string
        - in: header
          name: api_key
          description: ''
          required: true
          type: string
      responses:
        '200':
          description: Operación exitosa
          schema:
            $ref: '#/definitions/horariosEsta'
        '400':
          description: Solicitud incorrecta
          schema:
            $ref: '#/definitions/mensajeModel'
        '403':
          description: Credenciales incorrectas
          schema:
            $ref: '#/definitions/mensajeModel'
        '500':
          description: Error en sistema
          schema:
            $ref: '#/definitions/mensajeModel'
      security:
        - api_key: []
  /estacionamiento/{id}:
    delete:
      tags:
        - Estacionamiento
        - User
      x-swagger-router-controller: estacionamiento
      operationId: deleteEstacionamiento
      summary: Elimina el estacionamiento del usuario
      description: ''
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          description: ID del estacionamiento el cual sera eliminado
          required: true
          type: string
        - in: header
          name: api_key
          description: ''
          required: true
          type: string
      responses:
        '200':
          description: Operación exitosa
          schema:
            $ref: '#/definitions/mensajeModel'
        '400':
          description: Solicitud incorrecta
          schema:
            $ref: '#/definitions/mensajeModel'
        '403':
          description: Credenciales incorrectas
          schema:
            $ref: '#/definitions/mensajeModel'
        '500':
          description: Error en sistema
          schema:
            $ref: '#/definitions/mensajeModel'
      security:
        - api_key: []
  /estacionamiento/geo:
    get:
      tags:
        - Estacionamiento
      x-swagger-router-controller: estacionamiento
      operationId: findByGeo
      summary: Busca un estacionamiento  dada la geolocalicación
      description: ''
      produces:
        - application/json
      parameters:
        - in: query 
          name: lat
          description: Latitud del la ubicación.
          required: true
          type: number
        - in: query
          name: lon
          description: Longitud del la ubicación.
          required: true
          type: number
        - in: query
          name: rango
          description: Distancia desde la ubicación.
          type: string
        - in: header
          name: api_key
          description: ''
          required: true
          type: string
      responses:
        '200':
          description: Operación exitosa
          schema:
            $ref: '#/definitions/estacionamientos'
        '400':
          description: Solicitud incorrecta
          schema:
            $ref: '#/definitions/mensajeModel'
        '403':
          description: Credenciales incorrectas
          schema:
            $ref: '#/definitions/mensajeModel'
        '500':
          description: Error en sistema
          schema:
            $ref: '#/definitions/mensajeModel'
      security:
        - api_key: []
  /estacionamiento/disabled:
    post:
      tags:
        - Estacionamiento
      x-swagger-router-controller: estacionamiento
      operationId: disabled
      summary: Desabilita un estacionamiento
      description: ''
      produces:
        - application/json
      parameters:
        - in: body
          name: id
          description: Id del estacionamiento
          required: true
          schema:
            type: number
            format: int64
        - in: header
          name: api_key
          description: ''
          required: true
          type: string
      responses:
        '200':
          description: Operación exitosa
          schema:
            $ref: '#/definitions/mensajeModel'
        '400':
          description: Solicitud incorrecta
          schema:
            $ref: '#/definitions/mensajeModel'
        '403':
          description: Credenciales incorrectas
          schema:
            $ref: '#/definitions/mensajeModel'
        '500':
          description: Error en sistema
          schema:
            $ref: '#/definitions/mensajeModel'
      security:
        - api_key: []
  /estacionamiento/enabled:
    post:
      tags:
        - Estacionamiento
      x-swagger-router-controller: estacionamiento
      operationId: enabled
      summary: Habilita un estacionamiento
      description: ''
      produces:
        - application/json
      parameters:
        - in: body
          name: id
          description: Id del estacionamiento
          required: true
          schema:
            type: number
            format: int64
        - in: header
          name: api_key
          description: ''
          required: true
          type: string
      responses:
        '200':
          description: Operación exitosa
          schema:
            $ref: '#/definitions/mensajeModel'
        '400':
          description: Solicitud incorrecta
          schema:
            $ref: '#/definitions/mensajeModel'
        '403':
          description: Credenciales incorrectas
          schema:
            $ref: '#/definitions/mensajeModel'
        '500':
          description: Error en sistema
          schema:
            $ref: '#/definitions/mensajeModel'
      security:
        - api_key: []
  /estacionamiento/comuna:
    get:
      tags:
        - Estacionamiento
      x-swagger-router-controller: estacionamiento
      operationId: findByComuna
      summary: Busca un estacionamiento  dada la comuna
      description: ''
      produces:
        - application/json
      parameters:
        - in: query 
          name: comuna
          description: Id de la comuna.
          required: true
          type: number
        - in: header
          name: api_key
          description: ''
          required: true
          type: string
      responses:
        '200':
          description: Operación exitosa
          schema:
            $ref: '#/definitions/estacionamientos'
        '400':
          description: Solicitud incorrecta
          schema:
            $ref: '#/definitions/mensajeModel'
        '403':
          description: Credenciales incorrectas
          schema:
            $ref: '#/definitions/mensajeModel'
        '500':
          description: Error en sistema
          schema:
            $ref: '#/definitions/mensajeModel'
      security:
        - api_key: []
...
definitions:
  mensajeModel:
    type: object
    properties:
      message:
        type: string
      code:
        type: string
  credenciales:
    type: object
    properties:
      correo:
        type: string
      password:
        type: string
...
  userInfo:
    type: object
    properties:
      rut:
        type: integer
        format: int64
      div:
        type: string
      nombre:
        type: string
      apPaterno:
        type: string
      apMaterno:
        type: string
      correo:
        type: string
      telefono:
        type: integer
        format: int64
      estado:
        type: integer
        format: int64     
...
  dataLoginAdmin:
    type: object
    properties:
      usuarios:
        $ref: '#/definitions/usuarios'
      token:
        type: string
  dataLoginAdminF:
    type: object
    properties:
      nombre:
        type: string
      apPaterno:
        type: string
      apMaterno:
        type: string
      correo:
        type: string
      token:
        type: string
  estadoPagos:
    type: object
    properties:
      total:
        type: integer
        format: int64
      aprovados:
        type: integer
        format: int64
      reprovados:
        type: integer
        format: int64
      pendientes:
        type: integer
        format: int64
  usuarios:
    type: array
    items:
      $ref: '#/definitions/userInfo'
  horariosEsta:
    type: array
    items:
      $ref: '#/definitions/horario'
...
  auto:
    type: object
    properties:
      patente:
        type: string
      marca:
        type: string
      color:
        type: string
      motor:
        type: string
      chasis:
        type: string
  estacionamientos:
    type: array
    items:
      $ref: '#/definitions/estacionamientoFull'
  estacionamiento:
    type: object
    properties:
      id:
        description: Id del estacionamiento, la cual no se envia al momento de crear un estacionamiento
        type: integer
        format: int64
      horarios:
        description: indica la cantidad de horararios que tiene activos el estacionamiento
        type: integer
        format: int64
      costo:
        type: integer
        format: int64
      geo:
        $ref: '#/definitions/geo'
      comunaId:
        type: integer
        format: int64
      estado:
        type: integer
        format: int64
...
  estacionamientoFull:
    type: object
    properties:
      costo:
        type: integer
        format: int64
      geo:
        $ref: '#/definitions/geo'
      comunaId:
        type: integer
        format: int64
      comuna:
        type: string
      id:
        type: integer
        format: int64
      dueño:
        $ref: '#/definitions/dueño'
...
  pagosUsuarios:
    type: array
    items:
      $ref: '#/definitions/detallePago' 
...
  dueño:
    type: object
    properties:
      nombre:
        type: string
      apPaterno:
        type: string
      apMaterno:
        type: string
      rut:
        type: integer
        format: int64
      telefono:
        type: integer
        format: int64
      correo:
        type: string
  horario:
    type: object
    properties:
      fecha:
        type: string
        format: dateTime
      horaTermino:
        type: string
        format: dateTime
      horaInicio:
        type: string
        format: dateTime
      idEstacionamiento:
        type: integer
        format: int64
      idHorario:
        type: integer
        format: int64
      type:
        description: puede ser del tipo Ocupado o Libre
        type: string
      idReserva:
        description: si es del tipo ocupado devuelve el id de la reserva
        type: integer
        format: int64
...
  detallePago:
    type: object
    properties:
      id:
        type: integer
        format: int64
      rut:
        type: integer
        format: int64
      div:
        type: string
      nombre:
        type: string
      apPaterno:
        type: string
      apMaterno:
        type: string
      total:
        type: integer
        format: int64
      estadoPago:
        description: puede ser del tipo Aprovado  o rechazado
        type: string
      fechaPago:
        type: string
        format: dateTime
  reserva:
    type: object
    properties:
      id:
        type: integer
        format: int64
      horaEntrada:
        type: string
        format: dateTime
      horaSalida:
        type: string
        format: dateTime
      fecha:
        type: string
        format: dateTime
      total:
        type: integer
        format: int64
      autoPatente:
        type: string
      telefono:
        type: integer
        format: int64
      correo:
        type: string
      nombre:
        type: string
      apPaterno:
        type: string
      apMaterno:
        type: string
      idHorario:
        type: integer
        format: int64
  geo:
    type: object
    properties:
      lat:
        type: number
        format: float
      lon:
        type: number
        format: float
securityDefinitions:
  api_key:
    type: apiKey
    name: api_key
    in: header